build: added logic for delete label

+++ src/main/java/hexlet/code/app/service/LabelService.java
@@ -3,9 +3,11 @@ package hexlet.code.app.service;
 import hexlet.code.app.dto.LabelDTO.LabelCreateDTO;
 import hexlet.code.app.dto.LabelDTO.LabelDTO;
 import hexlet.code.app.dto.LabelDTO.LabelUpdateDTO;
+import hexlet.code.app.exception.MethodNotAllowedException;
 import hexlet.code.app.exception.ResourceNotFoundException;
 import hexlet.code.app.mapper.LabelMapper;
 import hexlet.code.app.repository.LabelRepository;
+import hexlet.code.app.repository.TaskRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,6 +21,9 @@ public class LabelService {
 
     @Autowired
     private LabelMapper labelMapper;
+
+    @Autowired
+    private TaskRepository taskRepository;
 
     public List<LabelDTO> getAllLabels() {
         return labelRepository.findAll().stream()
@@ -47,6 +52,11 @@ public class LabelService {
     }
 
     public void deleteLabel(Long labelId) {
+        var label = labelRepository.findById(labelId);
+
+        if (label.isPresent() && taskRepository.findByLabelsName(label.get().getName()).isPresent()) {
+            throw new MethodNotAllowedException("Label still has task");
+        }
         labelRepository.deleteById(labelId);
     }
